+++ apps/api/src/app/app.module.ts
@@ -1,8 +1,8 @@
-import { Module } from '@nestjs/common'
+import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo'
+import { Logger, Module } from '@nestjs/common'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { GraphQLModule } from '@nestjs/graphql'
 import { MongooseModule } from '@nestjs/mongoose'
-import { ScheduleModule } from '@nestjs/schedule'
 
 import { GraphQLError } from 'graphql'
 import { join } from 'path'
@@ -27,15 +27,17 @@ import { AppService } from './app.service'
       load: [configuration],
       envFilePath: ['.env', '.env.local'],
     }),
-    GraphQLModule.forRootAsync({
+    GraphQLModule.forRootAsync<ApolloDriverConfig>({
+      driver: ApolloDriver,
       useFactory: (configService: ConfigService) => ({
         // needed in production, to make apollo server work
         typePaths: [join(__dirname, '/../**/*.graphql')],
         definitions:
           configService.get<string>('env') === 'development'
             ? {
-                path: join(process.cwd(), 'src/graphql.ts'),
+                path: join(__dirname, '../graphql.ts'),
                 outputAs: 'class',
+                enumsAsTypes: true,
               }
             : null,
         playground: true,
@@ -45,17 +47,22 @@ import { AppService } from './app.service'
         },
         path: '/_api/graphql',
         context: ({ req, res }: GraphQLExpressContext) => ({ req, res }),
-        formatError: (error: GraphQLError) => {
-          const graphQLFormattedError = {
-            message: error?.extensions?.exception?.response?.message || error.message,
-            path: error.path,
-            locations: error.locations,
-            reason: error?.extensions?.exception?.response?.reason,
-            status: error?.extensions?.exception?.status,
-            exception: error?.extensions?.exception,
+        formatError: (formattedError, error: unknown) => {
+          if (error instanceof GraphQLError) {
+            const graphQLFormattedError = {
+              message: error.message,
+              path: error.path,
+              locations: error.locations,
+              extensions: {
+                code: error?.extensions?.code,
+              },
+            }
+            return graphQLFormattedError
           }
-          return graphQLFormattedError
+          return formattedError
         },
+        includeStacktraceInErrorResponses: false,
+        allowBatchedHttpRequests: true,
       }),
       inject: [ConfigService],
     }),
@@ -71,7 +78,6 @@ import { AppService } from './app.service'
     UserModule,
     AuthModule,
     ReviewModule,
-    ScheduleModule.forRoot(),
     ClientLoggingModule,
     OverrideModule,
     ComputationModule,