+++ apps/api/src/auth/auth.module.ts
@@ -3,8 +3,8 @@ import { ConfigService } from '@nestjs/config'
 import { JwtModule } from '@nestjs/jwt'
 import { MongooseModule } from '@nestjs/mongoose'
 
-import { RefreshToken, RefreshTokenSchema } from '../schemas/auth.schema'
-import { UserSchema } from '../schemas/user.schema'
+import { ModelName, RefreshTokenSchema, UserSchema } from '@cgr/schema'
+
 import { AnonymousStrategy } from './anonymous.strategy'
 import { AuthController } from './auth.controller'
 import { AuthService } from './auth.service'
@@ -13,8 +13,8 @@ import { JwtStrategy } from './jwt.strategy'
 @Module({
   imports: [
     MongooseModule.forFeature([
-      { name: 'user', schema: UserSchema },
-      { name: RefreshToken.name, schema: RefreshTokenSchema },
+      { name: ModelName.User, schema: UserSchema },
+      { name: ModelName.RefreshToken, schema: RefreshTokenSchema },
     ]),
     JwtModule.registerAsync({
       useFactory: async (configService: ConfigService) => ({