+++ apps/api/src/course/course.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common'
 import { MongooseModule } from '@nestjs/mongoose'
 
-import { CourseSchema } from '../schemas/course.schema'
+import { CourseSchema, ModelName } from '@cgr/schema'
+
 import { CourseResolver } from './course.resolver'
 import { CourseService } from './course.service'
 
 @Module({
-  imports: [MongooseModule.forFeature([{ name: 'course', schema: CourseSchema }])],
+  imports: [MongooseModule.forFeature([{ name: ModelName.Course, schema: CourseSchema }])],
   providers: [CourseResolver, CourseService],
   exports: [CourseService],
 })