+++ apps/api/src/override/override.resolver.ts
@@ -2,8 +2,7 @@ import { UseGuards } from '@nestjs/common'
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
 
 import { AdminAuthGuard } from '@api/auth/admin.guard'
-import { CourseGroupInput, OverrideInput } from '@api/graphql'
-import { Override } from '@api/schemas/override.schema'
+import { Override as GraphQLOverride, OverrideInput } from '@api/graphql'
 
 import { OverrideService } from './override.service'
 
@@ -12,7 +11,7 @@ export class OverrideResolver {
   constructor(private readonly overrideService: OverrideService) {}
 
   @Query('overrides')
-  getOverrides(): Promise<Override[]> {
+  getOverrides(): Promise<GraphQLOverride[]> {
     return this.overrideService.getOverrides()
   }
 
@@ -24,11 +23,7 @@ export class OverrideResolver {
 
   @UseGuards(AdminAuthGuard)
   @Mutation('deleteOverride')
-  deleteOverride(
-    @Args('courseNo') courseNo: string,
-    @Args('courseGroup') courseGroup: CourseGroupInput
-  ) {
-    const { studyProgram, academicYear, semester } = courseGroup
-    return this.overrideService.deleteOverride(courseNo, studyProgram, academicYear, semester)
+  deleteOverride(@Args('courseNo') courseNo: string) {
+    return this.overrideService.deleteOverride(courseNo)
   }
 }