+++ apps/api/src/review/review.service.ts
@@ -1,27 +1,50 @@
-import {
-  BadRequestException,
-  ConflictException,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common'
+import { UserInputError } from '@nestjs/apollo'
+import { Injectable, Logger } from '@nestjs/common'
+import { ConfigService } from '@nestjs/config'
 import { InjectModel } from '@nestjs/mongoose'
 
-import { StudyProgram } from '@thinc-org/chula-courses'
+import { IncomingWebhook } from '@slack/webhook'
 import { Model, Types } from 'mongoose'
 
+import { BadRequestError, NotFoundError } from '@api/common/errors'
+
 import {
-  CreateReviewInput,
-  EditReviewInput,
-  StudyProgram as GraphQLStudyProgram,
   Review,
+  ReviewDocument,
   ReviewInteractionType,
   ReviewStatus,
-} from '../graphql'
-import { ReviewDocument } from '../schemas/review.schema'
+  StudyProgram,
+} from '@cgr/schema'
+
+import { CreateReviewInput, EditReviewInput } from '../graphql'
 
 @Injectable()
 export class ReviewService {
-  constructor(@InjectModel('review') private reviewModel: Model<ReviewDocument>) {}
+  private logger: Logger = new Logger('ReviewService')
+  private webhook: IncomingWebhook
+
+  constructor(
+    private configService: ConfigService,
+    @InjectModel('review') private reviewModel: Model<Review>
+  ) {
+    const env = this.configService.get<string>('env')
+    const url = this.configService.get<string>('slackWebhookUrl')
+    if (url && env == 'production') {
+      this.logger.log(`Slack webhook is configured: ${url}`)
+      this.webhook = new IncomingWebhook(url)
+    }
+  }
+
+  async sendReviewAlert(review: ReviewDocument) {
+    if (!this.webhook) {
+      return
+    }
+    const reviewDashboardUrl = this.configService.get<string>('reviewDashboardUrl')
+    this.logger.log(`sent alert`)
+    return this.webhook.send({
+      text: `A new review is created for course ${review.courseNo} ${review.studyProgram} ${review.semester}/${review.academicYear}. Review them now in <${reviewDashboardUrl}|Review Dashboard>.`,
+    })
+  }
 
   async getReviews(): Promise<ReviewDocument[]> {
     const reviews = await this.reviewModel.find()
@@ -31,12 +54,9 @@ export class ReviewService {
   async create(
     { courseNo, semester, academicYear, studyProgram, rating, content }: CreateReviewInput,
     userId: string
-  ): Promise<Review> {
+  ): Promise<ReviewDocument> {
     if (rating < 0 || rating > 10) {
-      throw new BadRequestException({
-        reason: 'RATING_OUT_OF_BOUND',
-        message: `Rating must be between 0 and 10. Got ${rating}`,
-      })
+      throw new UserInputError(`Rating must be between 0 and 10. Got ${rating}`)
     }
 
     const review = await this.reviewModel.findOne({
@@ -45,10 +65,7 @@ export class ReviewService {
       studyProgram,
     })
     if (review) {
-      throw new ConflictException({
-        reason: 'DUPLICATE_REVIEW',
-        message: 'User already created a review for this course.',
-      })
+      throw new BadRequestError('User already created a review for this course.')
     }
 
     const newReview = new this.reviewModel({
@@ -61,118 +78,103 @@ export class ReviewService {
       content,
       status: 'PENDING',
     })
-
-    return this.transformReview(await newReview.save(), userId)
+    await newReview.save()
+    await this.sendReviewAlert(newReview)
+    return newReview
   }
 
   async getApprovedReviews(
     courseNo: string,
     studyProgram: StudyProgram,
     userId: string
-  ): Promise<Review[]> {
+  ): Promise<ReviewDocument[]> {
     const reviews = await this.reviewModel.find({
       courseNo,
       studyProgram,
-      status: ReviewStatus.APPROVED,
+      status: 'APPROVED',
     })
-    return reviews
-      .map((rawReview) => this.transformReview(rawReview, userId))
-      .sort((reviewA, reviewB) => {
-        if (reviewA.isOwner) {
-          return -1
-        }
-        if (reviewB.isOwner) {
-          return 1
-        }
-        // Sort by _id descending
-        return reviewB._id.localeCompare(reviewA._id)
-      })
+    return reviews.sort((reviewA, reviewB) => {
+      if (reviewA.ownerId.equals(userId)) {
+        return -1
+      }
+      if (reviewB.ownerId.equals(userId)) {
+        return 1
+      }
+      // Sort by _id descending
+      return reviewB._id.toString().localeCompare(reviewA._id.toString())
+    })
   }
 
-  async getPending(): Promise<Review[]> {
+  async getPending(): Promise<ReviewDocument[]> {
     const reviews = await this.reviewModel.find({
-      status: ReviewStatus.PENDING,
+      status: 'PENDING',
     })
-    return reviews.map((rawReview) => this.transformReview(rawReview, null))
+    return reviews
   }
 
   async getPendingForUser(
     courseNo: string,
     studyProgram: StudyProgram,
     userId: string
-  ): Promise<Review[]> {
+  ): Promise<ReviewDocument[]> {
     const reviews = await this.reviewModel.find({
-      $or: [{ status: ReviewStatus.PENDING }, { status: ReviewStatus.REJECTED }],
+      $or: [{ status: 'PENDING' }, { status: 'REJECTED' }],
       ownerId: userId,
       courseNo,
       studyProgram,
     })
-    return reviews.map((rawReview) => this.transformReview(rawReview, userId))
+    return reviews
   }
 
   async editMyReview(
     reviewId: string,
     reviewInput: EditReviewInput,
     userId: string
-  ): Promise<Review> {
+  ): Promise<ReviewDocument> {
     if (reviewInput.rating != null && (reviewInput.rating < 0 || reviewInput.rating > 10)) {
-      throw new BadRequestException({
-        reason: 'RATING_OUT_OF_BOUND',
-        message: `Rating must be between 0 and 10. Got ${reviewInput.rating}`,
-      })
+      throw new UserInputError(`Rating must be between 0 and 10. Got ${reviewInput.rating}`)
     }
 
     const review = await this.reviewModel.findById(reviewId)
     if (!review) {
-      throw new NotFoundException({
-        reason: 'REVIEW_NOT_FOUND',
-        message: 'Review with the given id does not exist.',
-      })
+      throw new NotFoundError('Review with the given id does not exist.')
     }
     if (!review.ownerId.equals(userId)) {
-      throw new BadRequestException({
-        reason: 'INVALID_USER',
-        message: 'Only the owner of the review can edit it',
-      })
+      throw new BadRequestError('User is not owner of this review.')
     }
     const newReview = await this.reviewModel.findByIdAndUpdate(
       reviewId,
       {
         $set: {
           ...reviewInput,
-          status: ReviewStatus.PENDING,
+          status: 'PENDING',
         },
       },
       { new: true }
     )
-    return this.transformReview(newReview, userId)
+    return newReview
   }
 
-  async remove(reviewId: string, userId: string): Promise<Review> {
+  async remove(reviewId: string, userId: string): Promise<ReviewDocument> {
     const review = await this.reviewModel.findOneAndDelete({
       _id: reviewId,
       ownerId: userId,
     })
     if (!review) {
-      throw new NotFoundException({
-        reason: 'REVIEW_NOT_FOUND',
-        message: `Either the review does not exist or the user is not the owner of the review`,
-      })
+      throw new NotFoundError(
+        'Either the review does not exist or the user is not the owner of the review'
+      )
     }
-    return this.transformReview(review, userId)
+    return review
   }
 
-  // TODO: hide reviews?
   async setStatus(
     reviewId: string,
     status: ReviewStatus,
     rejectionReason: string = null
-  ): Promise<string> {
-    if (status !== ReviewStatus.APPROVED && status !== ReviewStatus.REJECTED) {
-      throw new BadRequestException({
-        reason: 'INVALID_STATUS',
-        message: 'Only APPROVED and REJECTED status is supported',
-      })
+  ): Promise<ReviewDocument> {
+    if (status !== 'APPROVED' && status !== 'REJECTED') {
+      throw new UserInputError(`Only APPROVED and REJECTED status is supported. Got ${status}`)
     }
     const review = await this.reviewModel.findByIdAndUpdate(reviewId, {
       $set: {
@@ -181,25 +183,19 @@ export class ReviewService {
       },
     })
     if (!review) {
-      throw new NotFoundException({
-        reason: 'REVIEW_NOT_FOUND',
-        message: `Error setting status for review ${reviewId}: Review not found`,
-      })
+      throw new NotFoundError('Review with the given id does not exist.')
     }
-    return 'Review status updated successfully'
+    return review
   }
 
   async setInteraction(
     reviewId: string,
     interaction: ReviewInteractionType,
     userId: string
-  ): Promise<Review> {
+  ): Promise<ReviewDocument> {
     const review = await this.reviewModel.findById(reviewId)
     if (!review) {
-      throw new NotFoundException({
-        reason: 'REVIEW_NOT_FOUND',
-        message: 'Review with the given id does not exist.',
-      })
+      throw new NotFoundError('Review with the given id does not exist.')
     }
 
     const index = review.interactions.findIndex((interaction) => interaction.userId.equals(userId))
@@ -215,31 +211,6 @@ export class ReviewService {
     }
 
     await review.save()
-    return this.transformReview(review, userId)
-  }
-
-  private transformReview(rawReview: ReviewDocument, userId: string): Review {
-    const likeCount = rawReview.interactions.filter(
-      (interaction) => interaction.type === 'L'
-    ).length
-    const dislikeCount = rawReview.interactions.length - likeCount
-    const interactionType = rawReview.interactions.find((interaction) =>
-      interaction.userId.equals(userId)
-    )?.type
-    return {
-      _id: rawReview._id.toString(),
-      rating: rawReview.rating,
-      courseNo: rawReview.courseNo,
-      semester: rawReview.semester,
-      academicYear: rawReview.academicYear,
-      studyProgram: rawReview.studyProgram as GraphQLStudyProgram,
-      content: rawReview.content,
-      likeCount: likeCount,
-      dislikeCount: dislikeCount,
-      myInteraction: interactionType,
-      status: rawReview.status,
-      rejectionReason: rawReview.rejectionReason,
-      isOwner: rawReview.ownerId.equals(userId),
-    }
+    return review
   }
 }