+++ apps/admin-web/CHANGELOG.md
@@ -1,5 +1,41 @@
 # admin-web
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/codegen@1.2.0
+  - @cgr/course-utils@1.0.2
+
+## 0.1.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/codegen@1.2.0-beta.0
+  - @cgr/course-utils@1.0.2-beta.0
+
+## 0.1.0
+
+### Minor Changes
+
+- e1b5095b: upgrade to node 18
+
+### Patch Changes
+
+- 15dfbfca: Disable linting during Next.js builds
+
+## 0.1.0-beta.1
+
+### Minor Changes
+
+- e1b5095b: upgrade to node 18
+
 ## 0.0.3-beta.0
 
 ### Patch Changes