+++ apps/api/CHANGELOG.md
@@ -1,5 +1,85 @@
 # api
 
+## 1.4.1
+
+### Patch Changes
+
+- 85f089b2: fix override module to support new schema
+- Updated dependencies [85f089b2]
+  - @cgr/schema@1.1.0
+
+## 1.4.1-beta.0
+
+### Patch Changes
+
+- 85f089b2: fix override module to support new schema
+- Updated dependencies [85f089b2]
+  - @cgr/schema@1.1.0-beta.0
+
+## 1.4.0
+
+### Minor Changes
+
+- 7be734a2: - migrate graphql from apollo-server-express to @apollo/server
+  - use common types from `@cgr/schema`
+  - refactor api to be consistent across modules
+  - throw `GraphQLError`s instead of `HttpException`s in resolvers and services
+  - validate graphql date parsing and serialization
+- 7be734a2: make exam period and period fields nullable in graphql
+
+### Patch Changes
+
+- 7be734a2: - add constant model names
+  - add CourseCartItem type
+- 7be734a2: remove calendarId stuffs and fix type in course cart
+- 7be734a2: upgrade dependencies
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/schema@1.0.0
+
+## 1.4.0-beta.0
+
+### Minor Changes
+
+- 7be734a2: - migrate graphql from apollo-server-express to @apollo/server
+  - use common types from `@cgr/schema`
+  - refactor api to be consistent across modules
+  - throw `GraphQLError`s instead of `HttpException`s in resolvers and services
+  - validate graphql date parsing and serialization
+- 7be734a2: make exam period and period fields nullable in graphql
+
+### Patch Changes
+
+- 7be734a2: - add constant model names
+  - add CourseCartItem type
+- 7be734a2: remove calendarId stuffs and fix type in course cart
+- 7be734a2: upgrade dependencies
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/schema@1.0.0-beta.0
+
+## 1.3.0
+
+### Minor Changes
+
+- 21122935: fix stack field in log file being array instead of string causing elastic to failed parsing it
+- e1b5095b: upgrade to node 18
+
+### Patch Changes
+
+- 76659ecb: send review alert on create review
+
+## 1.3.0-beta.0
+
+### Minor Changes
+
+- 21122935: fix stack field in log file being array instead of string causing elastic to failed parsing it
+- e1b5095b: upgrade to node 18
+
+### Patch Changes
+
+- 76659ecb: send review alert on create review
+
 ## 1.2.0
 
 ### Minor Changes