+++ apps/web/CHANGELOG.md
@@ -1,5 +1,57 @@
 # web
 
+## 1.3.2
+
+### Patch Changes
+
+- 54265779: add 2565/ฤดูร้อน semester
+
+## 1.3.2-beta.0
+
+### Patch Changes
+
+- 54265779: add 2565/ฤดูร้อน semester
+
+## 1.3.1
+
+### Patch Changes
+
+- 7be734a2: make exam period and period fields nullable in graphql
+- 7be734a2: remove calendarId stuffs and fix type in course cart
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/codegen@1.2.0
+  - @cgr/course-utils@1.0.2
+
+## 1.3.1-beta.0
+
+### Patch Changes
+
+- 7be734a2: make exam period and period fields nullable in graphql
+- 7be734a2: remove calendarId stuffs and fix type in course cart
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/codegen@1.2.0-beta.0
+  - @cgr/course-utils@1.0.2-beta.0
+
+## 1.3.0
+
+### Minor Changes
+
+- e1b5095b: upgrade to node 18
+
+### Patch Changes
+
+- 15dfbfca: Disable linting during Next.js builds
+
+## 1.3.0-beta.1
+
+### Minor Changes
+
+- e1b5095b: upgrade to node 18
+
 ## 1.2.2-beta.0
 
 ### Patch Changes