+++ packages/course-utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cgr/course-utils
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/codegen@1.2.0
+
+## 1.0.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+- Updated dependencies [7be734a2]
+  - @cgr/codegen@1.2.0-beta.0
+
 ## 1.0.1
 
 ### Patch Changes